perf: reduce unnecesary dpi checks and renderer updates
When handling LOCATIONCHANGE events, we longer check for DPI updates on every single event, and
instead only check it when the monitor has changed. To account for the case where the DPI changes
without a monitor change or window movement, we handle the WM_DPICHANGED message instead.
Additionally, instead of updating the renderer on every monitor change, we first compare the correct
renderer settings with the actual renderer settings to determine whether it actually needs to be
updated. I'm honestly not 100% sure how to feel about this commit since it introduces a fair bit of
additional complexity, and it's also possible that there is a regression somewhere. Oh well.