fix: extend some V2 backend stability changes to Legacy backend
I originally thought the Legacy backend's ID2D1HwndRenderTarget automatically handled certain
scenarios, such as display adapter changes, but I was wrong. The reason it still worked in those
scenarios is because I handled D2DERR_RECREATE_TARGET, but that doesn't do everything perfectly. As
such, to improve code consistency and stability, this commit goes ahead and extends some of the
changes I made for the V2 backend to the Legacy backend.