fix: replace is_rect_visible with is_window_minimized to improve multi-monitor support
This was actually a long-standing bug regarding minimized windows that should finally be fixed. The
reason is_rect_visible() was added in the first place was to deal with minimized windows where the
rect would be far offscreen (all four points would be near -32000), and accidentally updating and
rendering with this window rect would lead to an oddly sized border. And, due to the asynchronous
nature of SetWindowPos/ShowWindow, that oddly sized border sometimes showed up briefly regardless of
when the call to show/hide the border window was made even though I would always try to re-render
with the correct size. Thus, new window rects were checked using is_rect_visible() to ensure the
oddly sized border never rendered in the first place. This method worked fine on single monitor
setups, but it led to issues on multi-monitor setups when the primary display wasn't at the top-left
(so the point (0, 0) would be somewhere in the middle of the virtual screenspace, and it would be
normal to have a window rect with negative coordinates even when the window was fully visible). To
fix this, the checks using is_rect_visible() have been replaced with checks using
is_window_minimized(). Additionally, it's worth noting that there are now duplicated
is_window_visible/is_window_cloaked/is_window_minimized calls in the WM_APP_SHOWUNCLOAKED and
WM_APP_MINIMIZEEND window messages (once from the explicit calls within the message handling body
itself, and another within the should_show_border() call). I tried to find a good way to refactor
this part out, but I couldn't find a good way to do so whilst keeping the clarity of the function
names intact, so it is what it is.